ParksList routes & curl scripts for create, show
Additionally, registers parksList routes in server.js
Copies in CRUD routes from a previous project
and modifies create and show to fit parksList routes.
Create is an authenticated-route post request
that saves an array of park codes to a user's parksList.
Create also saves the parksList _id to the User.userList.
Will need to refactor this patter
so that routes will follow the single responsibility principal.

Show returns the user's parksList.