Adds route to retrieve park data from the NPS api
The exploreParks/:id route queries MongoDB for the parksList
associated with the given _id.
The parksList array is turned into a string and added to the url
for the NPS parks endpoint, which contains useful individual park data.
The &fields=images query requests the images field to be returned by
the NPS api in addition to the default response data.
Also requires fetch to allow for fetch call.