Revises exploreParks routes, resource route names
exploreParks route handles users without favorites parks, or who aren't
signed in (they will pass '0' as a value for :id).
If the :id value is '0', the route will query the NPS api
with the park codes in defaultParks. Otherwise, the user's favoriteParks
list is used to query the NPS api. If the user's favoriteParks list has
fewer than 10 items, then some of the defaultParks are added to the NPS
api query.