Adds UpdateOne Route for toggling favoriteParks
This route is passed a parkCode associated with the park that a user
initiated the api call from. The route checks to see if the code already
exists in the user's favoriteParks.list, and if true it removes the
parkCode from the list. If false, it adds the parkCode to the list.

This "update" route uses dot notation and favoriteParks.save() instead
of favoriteParks.update() because the client response needed to include
the complete favoriteParks.list. Model.update() returns the number
of modified documents, instead of the list.