Refactors Api return data and naming patterns
The api /favoriteParks endpoints now only return either
favoriteParksData, favoriteParksId or both. All endpoints also hit the
NPS api using the getParkData method, so that favoriteParksData can be
returned instead of the parkCodes that were original returned. Although
this means the NPS api is pinged more often, the data returned to the
client is much more useful to the user. This commit also standardizes
the type of data being sent to the client, allowing for more predictable
data flow.