Returns res.data === null if user has no parks
Fixes design bug where api would query the NPS api if a user made a
request despite not having any parks in their favoriteParks list.
Now if a user makes a request, an object with a data attribute set to
null is returned. This allows the client to handle null data the same
way it works with NPS api data.